introduce swift-java jextract subcommand, completing our migration
