+++ Samples/SwiftAndJavaJarSampleLib/Package.swift
@@ -43,7 +43,7 @@ let javaIncludePath = "\(javaHome)/include"
 let package = Package(
   name: "SwiftAndJavaJarSampleLib",
   platforms: [
-    .macOS(.v10_15)
+    .macOS(.v15)
   ],
   products: [
     .library(
