+++ Sources/SwiftJavaTool/Commands/JExtractCommand.swift
@@ -1,0 +1,94 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2024 Apple Inc. and the Swift.org project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of Swift.org project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+import Foundation
+import ArgumentParser
+import SwiftJavaLib
+import JavaKit
+import JavaKitJar
+import SwiftJavaLib
+import JExtractSwiftLib
+import JavaKitConfigurationShared
+
+/// Extract Java bindings from Swift sources or interface files.
+///
+/// Example usage:
+/// ```
+/// > swift-java jextract
+//      --input-swift Sources/SwiftyBusiness \
+///     --output-swift .build/.../outputs/SwiftyBusiness \
+///     --output-Java .build/.../outputs/Java
+/// ```
+extension SwiftJava {
+
+  struct JExtractCommand: SwiftJavaBaseAsyncParsableCommand, HasCommonOptions {
+    static let configuration = CommandConfiguration(
+      commandName: "jextract", // TODO: wrap-swift?
+      abstract: "Wrap Swift functions and types with Java bindings, making them available to be called from Java")
+
+    @OptionGroup var commonOptions: SwiftJava.CommonOptions
+
+    @Option(help: "The mode of generation to use for the output files. Used with jextract mode.")
+    var mode: JExtractGenerationMode = .ffm
+
+    @Option(help: "The name of the Swift module into which the resulting Swift types will be generated.")
+    var swiftModule: String
+
+    var effectiveSwiftModule: String {
+      swiftModule
+    }
+
+    @Option(help: "The Java package the generated Java code should be emitted into.")
+    var javaPackage: String? = nil
+
+    @Option(help: "The directory where generated Swift files should be written. Generally used with jextract mode.")
+    var outputSwift: String
+
+    @Option(help: "The directory where generated Java files should be written. Generally used with jextract mode.")
+    var outputJava: String
+  }
+}
+
+extension SwiftJava.JExtractCommand {
+  func runSwiftJavaCommand(config: inout Configuration) async throws {
+    if let javaPackage {
+      config.javaPackage = javaPackage
+    }
+    config.swiftModule = self.effectiveSwiftModule
+    config.outputJavaDirectory = outputJava
+    config.outputSwiftDirectory = outputSwift
+
+    if let inputSwift = commonOptions.inputSwift {
+      config.inputSwiftDirectory = inputSwift
+    } else if let swiftModule = config.swiftModule {
+      // This is a "good guess" technically a target can be somewhere else, but then you can use --input-swift
+      config.inputSwiftDirectory = "\(FileManager.default.currentDirectoryPath)/Sources/\(swiftModule)"
+    }
+
+    print("[debug][swift-java] Running swift-java in mode: " + "\(self.mode)".bold)
+
+    try jextractSwift(config: config)
+  }
+}
+
+extension SwiftJava.JExtractCommand {
+  func jextractSwift(
+    config: Configuration
+  ) throws {
+    try SwiftToJava(config: config).run()
+  }
+
+}
+
+extension JExtractGenerationMode: ExpressibleByArgument {}
